+++ packages/config/src/env.d.ts
@@ -1,7 +1,1 @@
-interface ImportMetaEnv {
-  readonly VITE_HULL_HOST?: string;
-}
-
-interface ImportMeta {
-  readonly env: ImportMetaEnv;
-}
+export {};